rename to SmaugLab
