+++ .github/workflows/publish.yml
@@ -100,7 +100,7 @@ jobs:
           import glob, sys, zipfile
 
           names = zipfile.ZipFile(glob.glob("dist/*.whl")[0]).namelist()
-          if "auglab/configs/transform_params_gpu.json" not in names:
+          if "smauglab/configs/transform_params_gpu.json" not in names:
               sys.exit("refusing to publish: wheel has no config JSONs")
           if len([n for n in names if n.endswith(".py")]) < 20:
               sys.exit("refusing to publish: wheel is missing modules")
@@ -112,6 +112,6 @@ jobs:
         env:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
-        # Uploads the sdist as well as the wheel; auglab currently has no sdist
+        # Uploads the sdist as well as the wheel; smauglab currently has no sdist
         # on PyPI, which blocks anyone who needs to build from source.
         run: twine upload dist/*
