+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
-# Contributing to AugLab
+# Contributing to SmaugLab
 
 Thanks for contributing. This page covers the development setup, the checks CI
 runs, and how versioning and releases work.
@@ -6,8 +6,8 @@ runs, and how versioning and releases work.
 ## Development setup
 
 ```bash
-git clone git@github.com:neuropoly/AugLab.git
-cd AugLab
+git clone git@github.com:neuropoly/SmaugLab.git
+cd SmaugLab
 
 python3 -m venv venv
 source venv/bin/activate
@@ -41,8 +41,8 @@ data on disk, so it is fast enough to gate every pull request.
 
 | File | What it covers |
 | --- | --- |
-| `helpers.py` | `AugLabTestCase` base class (RNG seeding, test volumes) and config lookup |
-| `test_imports.py` | Every module under `auglab/` imports cleanly |
+| `helpers.py` | `SmaugLabTestCase` base class (RNG seeding, test volumes) and config lookup |
+| `test_imports.py` | Every module under `smauglab/` imports cleanly |
 | `test_configs.py` | Every shipped config parses, builds a pipeline, runs a forward pass, and is reproducible under a fixed seed |
 | `test_transforms_gpu.py` | Each GPU transform in isolation |
 | `test_packaging.py` | Builds the real wheel and checks its contents |
@@ -57,13 +57,13 @@ python -m unittest discover -s unit_tests -t .
 Cases that vary over configs or transforms use `subTest`, so one bad config
 does not hide the rest and the failure names the offending item — look for
 `SUBFAILED(config=...)` in the output. Derive new test classes from
-`AugLabTestCase` to get seeded RNGs and the shared `tiny_volume()` /
+`SmaugLabTestCase` to get seeded RNGs and the shared `tiny_volume()` /
 `tiny_seg()` helpers.
 
 Transforms in `test_transforms_gpu.py` are discovered by introspection, so a
 new transform class is covered as soon as it lands — as long as it can be built
 with default arguments. If yours needs configuration, cover it by adding a
-config JSON under `auglab/configs/`, which `test_configs.py` picks up
+config JSON under `smauglab/configs/`, which `test_configs.py` picks up
 automatically.
 
 Note that these are smoke and contract tests: they check that transforms run,
@@ -103,7 +103,7 @@ rest.
 
 ## Dependency pins
 
-`kornia` is capped at `>=0.7.3,<0.9`. AugLab subclasses kornia's *private*
+`kornia` is capped at `>=0.7.3,<0.9`. SmaugLab subclasses kornia's *private*
 augmentation internals (`_AugmentationBase`, `RigidAffineAugmentationBase3D`,
 `augmentation.container.ops`, `_adapted_rsampling`, `_tuple_range_reader`),
 which move between minor releases — 0.8.3 removed `kornia.core.Module` and the
@@ -111,6 +111,6 @@ whole `kornia.utils.helpers` module. The `kornia-compat` CI job runs the suite
 against both ends of the supported range, so a break shows up here rather than
 in a user's training run.
 
-`auglab/transforms/gpu/contrast.py` imports the private
+`smauglab/transforms/gpu/contrast.py` imports the private
 `torchvision.transforms._functional_tensor`. It still exists as of torchvision
 0.28, but carries the same risk.
