+++ smauglab/trainers/nnUNetTrainerDAExt.py
@@ -23,10 +23,10 @@ from nnunetv2.training.nnUNetTrainer.nnUNetTrainer import nnUNetTrainer
 from nnunetv2.utilities.helpers import dummy_context
 from torch import autocast
 
-from auglab import configs
-from auglab.trainers.utils import DownsampleSegForDSTransformCustom
-from auglab.transforms.cpu.transforms import AugTransforms
-from auglab.transforms.gpu.transforms import AugTransformsGPU
+from smauglab import configs
+from smauglab.trainers.utils import DownsampleSegForDSTransformCustom
+from smauglab.transforms.cpu.transforms import AugTransforms
+from smauglab.transforms.gpu.transforms import AugTransformsGPU
 
 
 class nnUNetTrainerDAExt(nnUNetTrainer):
@@ -61,7 +61,7 @@ class nnUNetTrainerDAExt(nnUNetTrainer):
         ### Adds transforms
         # Load transform parameters from json file
         configs_path = importlib.resources.files(configs)
-        json_path = os.environ.get("AUGLAB_PARAMS_CPU_JSON", str(configs_path / "transform_params.json"))
+        json_path = os.environ.get("SMAUGLAB_PARAMS_CPU_JSON", str(configs_path / "transform_params.json"))
         transforms.append(
             AugTransforms(
                 json_path=json_path,
@@ -155,9 +155,9 @@ class nnUNetTrainerDAExtGPU(nnUNetTrainer):
 
         # Load transform parameters from json file
         configs_path = importlib.resources.files(configs)
-        json_path = os.environ.get("AUGLAB_PARAMS_GPU_JSON", str(configs_path / "transform_params_gpu.json"))
+        json_path = os.environ.get("SMAUGLAB_PARAMS_GPU_JSON", str(configs_path / "transform_params_gpu.json"))
         self.transforms = AugTransformsGPU(json_path=json_path).to(self.device)
-        print(f"Using AugLab GPU transforms with parameters from: {json_path}")
+        print(f"Using SmaugLab GPU transforms with parameters from: {json_path}")
 
         # Copy json transfrom parameters to output folder
         shutil.copy(json_path, os.path.join(self.output_folder, "transform_params_gpu_used_for_training.json"))
@@ -188,7 +188,7 @@ class nnUNetTrainerDAExtGPU(nnUNetTrainer):
         transforms = []
 
         configs_path = importlib.resources.files(configs)
-        json_path = os.environ.get("AUGLAB_PARAMS_GPU_JSON", str(configs_path / "transform_params_gpu.json"))
+        json_path = os.environ.get("SMAUGLAB_PARAMS_GPU_JSON", str(configs_path / "transform_params_gpu.json"))
         with open(json_path) as f:
             config = json.load(f)
 
@@ -353,9 +353,9 @@ class nnUNetTrainerDAExtHybrid(nnUNetTrainer):
 
         # Load transform parameters from json file
         configs_path = importlib.resources.files(configs)
-        json_path = os.environ.get("AUGLAB_PARAMS_HYBRID_JSON", str(configs_path / "transform_params_hybrid.json"))
+        json_path = os.environ.get("SMAUGLAB_PARAMS_HYBRID_JSON", str(configs_path / "transform_params_hybrid.json"))
         self.transforms = AugTransformsGPU(json_path=json_path).to(self.device)
-        print(f"Using AugLab hybrid transforms with parameters from: {json_path}")
+        print(f"Using SmaugLab hybrid transforms with parameters from: {json_path}")
 
         # Copy json transfrom parameters to output folder
         shutil.copy(json_path, os.path.join(self.output_folder, "transform_params_hybrid_used_for_training.json"))
@@ -388,7 +388,7 @@ class nnUNetTrainerDAExtHybrid(nnUNetTrainer):
         ### Adds transforms
         # Load transform parameters from json file
         configs_path = importlib.resources.files(configs)
-        json_path = os.environ.get("AUGLAB_PARAMS_HYBRID_JSON", str(configs_path / "transform_params_hybrid.json"))
+        json_path = os.environ.get("SMAUGLAB_PARAMS_HYBRID_JSON", str(configs_path / "transform_params_hybrid.json"))
         transforms.append(
             AugTransforms(
                 json_path=json_path,
