+++ smauglab/transforms/cpu/artifact.py
