+++ smauglab/transforms/gpu/transforms.py
@@ -6,8 +6,8 @@ import numpy as np
 import torch
 from torch import Tensor, nn
 
-from auglab.transforms.gpu.base import AugmentationSequentialCustom, ImageOnlyTransform
-from auglab.transforms.gpu.contrast import (
+from smauglab.transforms.gpu.base import AugmentationSequentialCustom, ImageOnlyTransform
+from smauglab.transforms.gpu.contrast import (
     RandomBiasFieldGPU,
     RandomBrightnessGPU,
     RandomClampGPU,
@@ -20,11 +20,19 @@ from auglab.transforms.gpu.contrast import (
     RandomInverseGPU,
     ZscoreNormalizationGPU,
 )
-from auglab.transforms.gpu.fromSeg import RandomRedistributeSegGPU
-from auglab.transforms.gpu.spatial import RandomAcqTransformGPU, RandomAffine3DCustom, RandomFlipTransformGPU, RandomLowResTransformGPU
+from smauglab.transforms.gpu.domain_transfer import RandomDomainTransferGPU
+from smauglab.transforms.gpu.fromSeg import RandomPALETTEGPU, RandomRedistributeSegGPU
+from smauglab.transforms.gpu.spatial import (
+    RandomAcqTransformGPU,
+    RandomAffine3DCustom,
+    RandomCropTransformGPU,
+    RandomFlipTransformGPU,
+    RandomLowResTransformGPU,
+)
+from smauglab.transforms.synthseg.transforms import RandomSynthSegGPU
 
 
-class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
+class AugTransformsGPU(AugmentationSequentialCustom):
     """
     Module to perform data augmentation on GPU.
     """
@@ -74,12 +82,74 @@ class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
                 )
             )
 
+        # SynthSeg generative augmentation: replace the image with a GMM synthesis
+        # of the segmentation (intensity-only here, so the mask stays consistent;
+        # geometric transforms above deform the labels first). All SynthSeg
+        # generator parameters are read straight from the config block.
+        synthseg_params = self.transform_params.get("SynthSeg")
+        if synthseg_params is not None:
+            synthseg_kwargs = {k: v for k, v in synthseg_params.items() if k != "probability"}
+            transforms.append(
+                RandomSynthSegGPU(
+                    p=synthseg_params.get("probability", 1.0),
+                    **synthseg_kwargs,
+                )
+            )
+
         ## Transfer augmentations (TA)
-        ta_transforms = []
+        #########################
+        # Replace image with V26_6_2 contrast (K-means + Voronoi + per-label remap)
+        palette_params = self.transform_params.get("RandomPALETTETransform")
+        if palette_params is not None:
+            transforms.append(
+                RandomPALETTEGPU(
+                    p=palette_params.get("probability", 1.0),
+                    c_choices=palette_params.get("c_choices", [2, 3, 4, 5, 6]),
+                    s_choices=palette_params.get("s_choices", [2, 3, 4, 5, 6, 7, 8, 9, 10]),
+                    blur_sigmas=palette_params.get("blur_sigmas", [0.0, 0.0, 0.0, 0.3, 0.5, 0.8]),
+                    dark_threshold=palette_params.get("dark_threshold", 0.01),
+                    n_kmeans_subsample=palette_params.get("n_kmeans_subsample", 10000),
+                    skip_parcellation_prob=palette_params.get("skip_parcellation_prob", 0.10),
+                    skip_sub_parc_prob=palette_params.get("skip_sub_parc_prob", 0.40),
+                    alpha_magnitude_range=palette_params.get("alpha_magnitude_range", [0.5, 2.0]),
+                    label_remap_prob=palette_params.get("label_remap_prob", 0.5),
+                    min_label_voxels=palette_params.get("min_label_voxels", 4),
+                    label_classes=palette_params.get("label_classes", None),
+                )
+            )
+
+        # Domain transfer: randomly re-render the image as another sequence/cluster (TA)
+        # Accept either the class-name key or the descriptive key.
+        domain_params = self.transform_params.get("RandomDomainTransferGPU") or self.transform_params.get("DomainTransferTransform")
+        if domain_params is not None:
+            transforms.append(
+                RandomDomainTransferGPU(
+                    bank_path=domain_params.get("bank_path", None),
+                    source_label=domain_params["source_label"],
+                    targets=domain_params.get("targets", None),
+                    include_self=domain_params.get("include_self", False),
+                    any_source=domain_params.get("any_source", False),
+                    sigma=domain_params.get("sigma", 2.0),
+                    apply_to_channel=domain_params.get("apply_to_channel", [0]),
+                    zscore_io=domain_params.get("zscore_io", "auto"),
+                    pct=domain_params.get("pct", 1.0),
+                    blend_targets=domain_params.get("blend_targets", 1),
+                    blend_concentration=domain_params.get("blend_concentration", 1.0),
+                    p_class_mix=domain_params.get("p_class_mix", 0.0),
+                    bias_field_std=domain_params.get("bias_field_std", 0.0),
+                    bias_scale=domain_params.get("bias_scale", 0.03),
+                    p_spatial_mix=domain_params.get("p_spatial_mix", 0.0),
+                    spatial_mix_scale=domain_params.get("spatial_mix_scale", 0.03),
+                    spatial_mix_gain=domain_params.get("spatial_mix_gain", 3.0),
+                    p=domain_params.get("probability", 0.0),
+                    same_on_batch=domain_params.get("same_on_batch", False),
+                )
+            )
+
         # Inverse transform (max - pixel_value)
         inverse_params = self.transform_params.get("InverseTransform")
         if inverse_params is not None:
-            ta_transforms.append(
+            transforms.append(
                 RandomInverseGPU(
                     p=inverse_params.get("probability", 0),
                     in_seg=inverse_params.get("in_seg", 0.0),
@@ -93,7 +163,7 @@ class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
         # Histogram manipulations
         histo_params = self.transform_params.get("HistogramEqualizationTransform")
         if histo_params is not None:
-            ta_transforms.append(
+            transforms.append(
                 RandomHistogramEqualizationGPU(
                     p=histo_params.get("probability", 0),
                     in_seg=histo_params.get("in_seg", 0.0),
@@ -107,18 +177,20 @@ class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
         # Redistribute segmentation values transform
         redistribute_params = self.transform_params.get("RedistributeSegTransform")
         if redistribute_params is not None:
-            ta_transforms.append(
+            transforms.append(
                 RandomRedistributeSegGPU(
                     in_seg=redistribute_params.get("in_seg", 0.2),
                     retain_stats=redistribute_params.get("retain_stats", False),
                     p=redistribute_params.get("probability", 0),
+                    std_noise_range=redistribute_params.get("std_noise_range", [0.1, 0.3]),
+                    dilation_iterations_range=redistribute_params.get("dilation_iterations_range", [1, 3]),
                 )
             )
 
         # Scharr filter
         scharr_params = self.transform_params.get("ScharrTransform")
         if scharr_params is not None:
-            ta_transforms.append(
+            transforms.append(
                 RandomConvTransformGPU(
                     kernel_type=scharr_params.get("kernel_type", "Scharr"),
                     p=scharr_params.get("probability", 0),
@@ -134,7 +206,7 @@ class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
         # Unsharp masking
         unsharp_params = self.transform_params.get("UnsharpMaskTransform")
         if unsharp_params is not None:
-            ta_transforms.append(
+            transforms.append(
                 RandomConvTransformGPU(
                     kernel_type=unsharp_params.get("kernel_type", "UnsharpMask"),
                     p=unsharp_params.get("probability", 0),
@@ -150,7 +222,7 @@ class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
         # RandomConv transform
         randconv_params = self.transform_params.get("RandomConvTransform")
         if randconv_params is not None:
-            ta_transforms.append(
+            transforms.append(
                 RandomConvTransformGPU(
                     kernel_type=randconv_params.get("kernel_type", "RandConv"),
                     p=randconv_params.get("probability", 0),
@@ -162,377 +234,6 @@ class AugTransformsGPURandomOrder(AugmentationSequentialCustom):
                     mix_prob=randconv_params.get("mix_prob", 0.0),
                 )
             )
-
-        # Apply functions
-        func_list = [
-            lambda x: torch.log(1 + x),
-            torch.sqrt,
-            torch.sin,
-            torch.exp,
-            lambda x: 1 / (1 + torch.exp(-x)),
-        ]
-        function_params = self.transform_params.get("FunctionTransform")
-        if function_params is not None:
-            ta_transforms.extend(
-                RandomFunctionGPU(
-                    func=func,
-                    p=function_params.get("probability", 0),
-                    in_seg=function_params.get("in_seg", 0.0),
-                    out_seg=function_params.get("out_seg", 0.0),
-                    mix_in_out=function_params.get("mix_in_out", False),
-                    retain_stats=function_params.get("retain_stats", False),
-                )
-                for func in func_list
-            )
-
-        # Bias field artifact
-        bias_field_params = self.transform_params.get("BiasFieldTransform")
-        if bias_field_params is not None:
-            ta_transforms.append(
-                RandomBiasFieldGPU(
-                    p=bias_field_params.get("probability", 0),
-                    in_seg=bias_field_params.get("in_seg", 0.0),
-                    out_seg=bias_field_params.get("out_seg", 0.0),
-                    mix_in_out=bias_field_params.get("mix_in_out", False),
-                    retain_stats=bias_field_params.get("retain_stats", False),
-                    coefficients=bias_field_params.get("coefficients", 0.5),
-                )
-            )
-
-        ## General enhancement (GE)
-        ge_transforms = []
-        # Clamping transform
-        clamp_params = self.transform_params.get("ClampTransform")
-        if clamp_params is not None:
-            ge_transforms.append(
-                RandomClampGPU(
-                    max_clamp_amount=clamp_params.get("max_clamp_amount", 0.0),
-                    in_seg=clamp_params.get("in_seg", 0.0),
-                    out_seg=clamp_params.get("out_seg", 0.0),
-                    mix_in_out=clamp_params.get("mix_in_out", False),
-                    retain_stats=clamp_params.get("retain_stats", False),
-                    p=clamp_params.get("probability", 0),
-                )
-            )
-
-        # Noise transforms
-        noise_params = self.transform_params.get("GaussianNoiseTransform")
-        if noise_params is not None:
-            ge_transforms.append(
-                RandomGaussianNoiseGPU(
-                    mean=noise_params.get("mean", 0.0),
-                    std=noise_params.get("std", 1.0),
-                    in_seg=noise_params.get("in_seg", 0.0),
-                    out_seg=noise_params.get("out_seg", 0.0),
-                    mix_in_out=noise_params.get("mix_in_out", False),
-                    p=noise_params.get("probability", 0),
-                )
-            )
-
-        # Gaussian blur
-        gaussianblur_params = self.transform_params.get("GaussianBlurTransform")
-        if gaussianblur_params is not None:
-            ge_transforms.append(
-                RandomConvTransformGPU(
-                    kernel_type=gaussianblur_params.get("kernel_type", "GaussianBlur"),
-                    in_seg=gaussianblur_params.get("in_seg", 0.0),
-                    out_seg=gaussianblur_params.get("out_seg", 0.0),
-                    mix_in_out=gaussianblur_params.get("mix_in_out", False),
-                    p=gaussianblur_params.get("probability", 0),
-                    sigma=gaussianblur_params.get("sigma", 1.0),
-                )
-            )
-
-        # Brightness transforms
-        brightness_params = self.transform_params.get("BrightnessTransform")
-        if brightness_params is not None:
-            ge_transforms.append(
-                RandomBrightnessGPU(
-                    brightness_range=brightness_params.get("brightness_range", [0.5, 1.5]),
-                    in_seg=brightness_params.get("in_seg", 0.0),
-                    out_seg=brightness_params.get("out_seg", 0.0),
-                    mix_in_out=brightness_params.get("mix_in_out", False),
-                    p=brightness_params.get("probability", 0),
-                )
-            )
-
-        # Gamma transforms
-        gamma_params = self.transform_params.get("GammaTransform")
-        if gamma_params is not None:
-            ge_transforms.append(
-                RandomGammaGPU(
-                    gamma_range=gamma_params.get("gamma_range", [0.7, 1.5]),
-                    p=gamma_params.get("probability", 0),
-                    invert_image=False,
-                    in_seg=gamma_params.get("in_seg", 0.0),
-                    out_seg=gamma_params.get("out_seg", 0.0),
-                    mix_in_out=gamma_params.get("mix_in_out", False),
-                    retain_stats=gamma_params.get("retain_stats", False),
-                )
-            )
-
-        inv_gamma_params = self.transform_params.get("InvGammaTransform")
-        if inv_gamma_params is not None:
-            ge_transforms.append(
-                RandomGammaGPU(
-                    gamma_range=inv_gamma_params.get("gamma_range", [0.7, 1.5]),
-                    p=inv_gamma_params.get("probability", 0),
-                    in_seg=inv_gamma_params.get("in_seg", 0.0),
-                    out_seg=inv_gamma_params.get("out_seg", 0.0),
-                    mix_in_out=inv_gamma_params.get("mix_in_out", False),
-                    invert_image=True,
-                    retain_stats=inv_gamma_params.get("retain_stats", False),
-                )
-            )
-
-        # nnUNetV2 Contrast transforms
-        contrast_params = self.transform_params.get("ContrastTransform")
-        if contrast_params is not None:
-            ge_transforms.append(
-                RandomContrastGPU(
-                    contrast_range=contrast_params.get("contrast_range", [0.75, 1.25]),
-                    p=contrast_params.get("probability", 0),
-                    in_seg=contrast_params.get("in_seg", 0.0),
-                    out_seg=contrast_params.get("out_seg", 0.0),
-                    mix_in_out=contrast_params.get("mix_in_out", False),
-                    retain_stats=contrast_params.get("retain_stats", False),
-                )
-            )
-
-        # Shape transforms (Cropping and Simulating low resolution)
-        lowres_params = self.transform_params.get("SimulateLowResTransform")
-        if lowres_params is not None:
-            transforms.append(
-                RandomLowResTransformGPU(
-                    p=lowres_params.get("probability", 0),
-                    scale=lowres_params.get("scale", [0.3, 1.0]),
-                    crop=lowres_params.get("crop", [1.0, 1.0]),
-                    same_on_batch=lowres_params.get("same_on_batch", False),
-                )
-            )
-
-        acq_params = self.transform_params.get("AcqTransform")
-        if acq_params is not None:
-            ge_transforms.append(
-                RandomAcqTransformGPU(
-                    p=acq_params.get("probability", 0),
-                    scale=acq_params.get("scale", [0.3, 1.0]),
-                    crop=acq_params.get("crop", [1.0, 1.0]),
-                    one_dim=True,
-                    same_on_batch=acq_params.get("same_on_batch", False),
-                )
-            )
-
-        ## Random Z-score normalization
-        zscore_params = self.transform_params.get("ZscoreNormalizationTransform")
-        if zscore_params is not None:
-            ge_transforms.append(ZscoreNormalizationGPU(p=zscore_params.get("probability", 0)))
-
-        choose_x_params = self.transform_params.get("RandomChooseXTransforms")
-        transforms.append(
-            RandomChooseXTransformsGPU(
-                transforms_list=ta_transforms,
-                num_transforms=len(ta_transforms),
-                p=choose_x_params.get("ta_probability", 1.0),
-                random_order=choose_x_params.get("ta_random_order", True),
-            )
-        )
-        transforms.append(
-            RandomChooseXTransformsGPU(
-                transforms_list=ge_transforms,
-                num_transforms=len(ge_transforms),
-                p=choose_x_params.get("ge_probability", 1.0),
-                random_order=choose_x_params.get("ge_random_order", True),
-            )
-        )
-
-        return transforms
-
-
-class AugTransformsGPURandomOrderTA(AugmentationSequentialCustom):
-    """
-    Module to perform data augmentation on GPU.
-    """
-
-    def __init__(self, json_path: str):
-        # Load transform parameters from JSON
-        config_path = os.path.join(json_path)
-        with open(config_path) as f:
-            config = json.load(f)
-
-        if "GPU" in config.keys():
-            self.transform_params = config["GPU"]
-        else:
-            self.transform_params = config
-
-        transforms = self._build_transforms()
-        super().__init__(
-            *transforms, data_keys=["input", "mask"], same_on_batch=True
-        )  # Same_on_batch to ensure mask are aligned with images correctly (custom) see AugmentationSequentialOpsCustom in base.py
-
-    def _build_transforms(self) -> list[nn.Module]:
-        transforms = []
-
-        # Flipping transforms
-        flip_params = self.transform_params.get("FlipTransform")
-        if flip_params is not None:
-            transforms.append(
-                RandomFlipTransformGPU(
-                    flip_axis=flip_params.get("flip_axis", [0]),
-                    p=flip_params.get("probability", 0),
-                    same_on_batch=flip_params.get("same_on_batch", False),
-                    keepdim=flip_params.get("keepdim", True),
-                )
-            )
-
-        # Spatial transforms
-        affine_params = self.transform_params.get("AffineTransform")
-        if affine_params is not None:
-            transforms.append(
-                RandomAffine3DCustom(
-                    degrees=affine_params.get("degrees", 10),
-                    translate=affine_params.get("translate", [0.1, 0.1, 0.1]),
-                    scale=affine_params.get("scale", [0.9, 1.1]),
-                    shears=affine_params.get("shear", [-10, 10, -10, 10, -10, 10]),
-                    resample=affine_params.get("resample", "bilinear"),
-                    p=affine_params.get("probability", 0),
-                )
-            )
-
-        ## Transfer augmentations (TA)
-        ta_transforms = []
-        # Inverse transform (max - pixel_value)
-        inverse_params = self.transform_params.get("InverseTransform")
-        if inverse_params is not None:
-            ta_transforms.append(
-                RandomInverseGPU(
-                    p=inverse_params.get("probability", 0),
-                    in_seg=inverse_params.get("in_seg", 0.0),
-                    out_seg=inverse_params.get("out_seg", 0.0),
-                    mix_in_out=inverse_params.get("mix_in_out", False),
-                    mix_prob=inverse_params.get("mix_prob", 0.0),
-                    retain_stats=inverse_params.get("retain_stats", False),
-                )
-            )
-
-        # Histogram manipulations
-        histo_params = self.transform_params.get("HistogramEqualizationTransform")
-        if histo_params is not None:
-            ta_transforms.append(
-                RandomHistogramEqualizationGPU(
-                    p=histo_params.get("probability", 0),
-                    in_seg=histo_params.get("in_seg", 0.0),
-                    out_seg=histo_params.get("out_seg", 0.0),
-                    mix_in_out=histo_params.get("mix_in_out", False),
-                    mix_prob=histo_params.get("mix_prob", 0.0),
-                    retain_stats=histo_params.get("retain_stats", False),
-                )
-            )
-
-        # Redistribute segmentation values transform
-        redistribute_params = self.transform_params.get("RedistributeSegTransform")
-        if redistribute_params is not None:
-            ta_transforms.append(
-                RandomRedistributeSegGPU(
-                    in_seg=redistribute_params.get("in_seg", 0.2),
-                    retain_stats=redistribute_params.get("retain_stats", False),
-                    p=redistribute_params.get("probability", 0),
-                )
-            )
-
-        # Scharr filter
-        scharr_params = self.transform_params.get("ScharrTransform")
-        if scharr_params is not None:
-            ta_transforms.append(
-                RandomConvTransformGPU(
-                    kernel_type=scharr_params.get("kernel_type", "Scharr"),
-                    p=scharr_params.get("probability", 0),
-                    in_seg=scharr_params.get("in_seg", 0.0),
-                    out_seg=scharr_params.get("out_seg", 0.0),
-                    mix_in_out=scharr_params.get("mix_in_out", False),
-                    retain_stats=scharr_params.get("retain_stats", True),
-                    absolute=scharr_params.get("absolute", True),
-                    mix_prob=scharr_params.get("mix_prob", 0.0),
-                )
-            )
-
-        # Unsharp masking
-        unsharp_params = self.transform_params.get("UnsharpMaskTransform")
-        if unsharp_params is not None:
-            ta_transforms.append(
-                RandomConvTransformGPU(
-                    kernel_type=unsharp_params.get("kernel_type", "UnsharpMask"),
-                    p=unsharp_params.get("probability", 0),
-                    in_seg=unsharp_params.get("in_seg", 0.0),
-                    out_seg=unsharp_params.get("out_seg", 0.0),
-                    mix_in_out=unsharp_params.get("mix_in_out", False),
-                    sigma=unsharp_params.get("sigma", 1.0),
-                    unsharp_amount=unsharp_params.get("unsharp_amount", 1.5),
-                    mix_prob=unsharp_params.get("mix_prob", 0.0),
-                )
-            )
-
-        # RandomConv transform
-        randconv_params = self.transform_params.get("RandomConvTransform")
-        if randconv_params is not None:
-            ta_transforms.append(
-                RandomConvTransformGPU(
-                    kernel_type=randconv_params.get("kernel_type", "RandConv"),
-                    p=randconv_params.get("probability", 0),
-                    in_seg=randconv_params.get("in_seg", 0.0),
-                    out_seg=randconv_params.get("out_seg", 0.0),
-                    mix_in_out=randconv_params.get("mix_in_out", False),
-                    retain_stats=randconv_params.get("retain_stats", False),
-                    kernel_sizes=randconv_params.get("kernel_sizes", [1, 3, 5, 7]),
-                    mix_prob=randconv_params.get("mix_prob", 0.0),
-                )
-            )
-
-        # Apply functions
-        func_list = [
-            lambda x: torch.log(1 + x),
-            torch.sqrt,
-            torch.sin,
-            torch.exp,
-            lambda x: 1 / (1 + torch.exp(-x)),
-        ]
-        function_params = self.transform_params.get("FunctionTransform")
-        if function_params is not None:
-            ta_transforms.extend(
-                RandomFunctionGPU(
-                    func=func,
-                    p=function_params.get("probability", 0),
-                    in_seg=function_params.get("in_seg", 0.0),
-                    out_seg=function_params.get("out_seg", 0.0),
-                    mix_in_out=function_params.get("mix_in_out", False),
-                    retain_stats=function_params.get("retain_stats", False),
-                )
-                for func in func_list
-            )
-
-        # Bias field artifact
-        bias_field_params = self.transform_params.get("BiasFieldTransform")
-        if bias_field_params is not None:
-            ta_transforms.append(
-                RandomBiasFieldGPU(
-                    p=bias_field_params.get("probability", 0),
-                    in_seg=bias_field_params.get("in_seg", 0.0),
-                    out_seg=bias_field_params.get("out_seg", 0.0),
-                    mix_in_out=bias_field_params.get("mix_in_out", False),
-                    retain_stats=bias_field_params.get("retain_stats", False),
-                    coefficients=bias_field_params.get("coefficients", 0.5),
-                )
-            )
-
-        choose_x_params = self.transform_params.get("RandomChooseXTransforms")
-        transforms.append(
-            RandomChooseXTransformsGPU(
-                transforms_list=ta_transforms,
-                num_transforms=len(ta_transforms),
-                p=choose_x_params.get("ta_probability", 1.0),
-                random_order=False,
-            )
-        )
 
         ## General enhancement (GE)
         # Clamping transform
@@ -633,6 +334,28 @@ class AugTransformsGPURandomOrderTA(AugmentationSequentialCustom):
                 )
             )
 
+        # Apply functions
+        func_list = [
+            lambda x: torch.log(1 + x),
+            torch.sqrt,
+            torch.sin,
+            torch.exp,
+            lambda x: 1 / (1 + torch.exp(-x)),
+        ]
+        function_params = self.transform_params.get("FunctionTransform")
+        if function_params is not None:
+            transforms.extend(
+                RandomFunctionGPU(
+                    func=func,
+                    p=function_params.get("probability", 0),
+                    in_seg=function_params.get("in_seg", 0.0),
+                    out_seg=function_params.get("out_seg", 0.0),
+                    mix_in_out=function_params.get("mix_in_out", False),
+                    retain_stats=function_params.get("retain_stats", False),
+                )
+                for func in func_list
+            )
+
         # Shape transforms (Cropping and Simulating low resolution)
         lowres_params = self.transform_params.get("SimulateLowResTransform")
         if lowres_params is not None:
@@ -640,7 +363,6 @@ class AugTransformsGPURandomOrderTA(AugmentationSequentialCustom):
                 RandomLowResTransformGPU(
                     p=lowres_params.get("probability", 0),
                     scale=lowres_params.get("scale", [0.3, 1.0]),
-                    crop=lowres_params.get("crop", [1.0, 1.0]),
                     same_on_batch=lowres_params.get("same_on_batch", False),
                 )
             )
@@ -651,9 +373,33 @@ class AugTransformsGPURandomOrderTA(AugmentationSequentialCustom):
                 RandomAcqTransformGPU(
                     p=acq_params.get("probability", 0),
                     scale=acq_params.get("scale", [0.3, 1.0]),
-                    crop=acq_params.get("crop", [1.0, 1.0]),
                     one_dim=True,
                     same_on_batch=acq_params.get("same_on_batch", False),
+                )
+            )
+
+        crop_params = self.transform_params.get("CropTransform")
+        if crop_params is not None:
+            transforms.append(
+                RandomCropTransformGPU(
+                    p=crop_params.get("probability", 0),
+                    crop=crop_params.get("crop", [1.0, 1.0]),
+                    pos=crop_params.get("pos", [0.0, 1.0]),
+                    same_on_batch=acq_params.get("same_on_batch", False),
+                )
+            )
+
+        # Bias field artifact
+        bias_field_params = self.transform_params.get("BiasFieldTransform")
+        if bias_field_params is not None:
+            transforms.append(
+                RandomBiasFieldGPU(
+                    p=bias_field_params.get("probability", 0),
+                    in_seg=bias_field_params.get("in_seg", 0.0),
+                    out_seg=bias_field_params.get("out_seg", 0.0),
+                    mix_in_out=bias_field_params.get("mix_in_out", False),
+                    retain_stats=bias_field_params.get("retain_stats", False),
+                    coefficients=bias_field_params.get("coefficients", 0.5),
                 )
             )
 
@@ -686,7 +432,6 @@ class RandomChooseXTransformsGPU(ImageOnlyTransform):
         same_on_batch: bool = False,
         p: float = 1.0,
         keepdim: bool = True,
-        random_order: bool = True,
         **kwargs,
     ) -> None:
         super().__init__(p=p, same_on_batch=same_on_batch, keepdim=keepdim)
@@ -694,7 +439,6 @@ class RandomChooseXTransformsGPU(ImageOnlyTransform):
             raise ValueError(f"num_transforms must be a non-negative int. Got {num_transforms!r}.")
         self.transforms_list = nn.ModuleList(transforms_list)
         self.num_transforms = num_transforms
-        self.random_order = random_order
 
     def _apply_mix(self, x: Tensor, seg: Tensor | None) -> Tensor:
         if self.num_transforms == 0 or len(self.transforms_list) == 0:
@@ -702,10 +446,7 @@ class RandomChooseXTransformsGPU(ImageOnlyTransform):
 
         k = min(self.num_transforms, len(self.transforms_list))
         # sample without replacement
-        if self.random_order:
-            idx = torch.randperm(len(self.transforms_list), device=x.device)[:k]
-        else:
-            idx = torch.arange(len(self.transforms_list), device=x.device)[:k]
+        idx = torch.randperm(len(self.transforms_list), device=x.device)[:k]
 
         child_params: dict[str, Tensor] = {}
         if seg is not None:
@@ -767,31 +508,30 @@ if __name__ == "__main__":
     # Example usage
     import importlib
 
-    from auglab import configs
-    from auglab.transforms.gpu.transforms import AugTransformsGPU
-    from auglab.utils.image import Image, resample_nib
+    from smauglab import configs
+    from smauglab.utils.image import Image, resample_nib
 
     configs_path = importlib.resources.files(configs)
     json_path = configs_path / "transform_params_gpu.json"
     augmentor = AugTransformsGPU(json_path)
 
     # Load images and masks tensors
-    img_path = "/home/GRAMES.POLYMTL.CA/p118739/data_nvme_p118739/data/datasets/data-multi-subject/sub-amu02/anat/sub-amu02_T1w.nii.gz"
+    img_path = "/home/ge.polymtl.ca/p118739/data/datasets/data-multi-subject/sub-amu02/anat/sub-amu02_T1w.nii.gz"
     img = Image(img_path).change_orientation("RSP")
     img = resample_nib(img, new_size=[1, 1, 1], new_size_type="mm", interpolation="linear")
     img_tensor = torch.from_numpy(img.data.copy()).to(torch.float32)
 
-    seg_path = "/home/GRAMES.POLYMTL.CA/p118739/data_nvme_p118739/data/datasets/data-multi-subject/derivatives/labels/sub-amu02/anat/sub-amu02_T1w_label-spine_dseg.nii.gz"
+    seg_path = "/home/ge.polymtl.ca/p118739/data/datasets/data-multi-subject/derivatives/labels/sub-amu02/anat/sub-amu02_T1w_label-spine_dseg.nii.gz"
     seg = Image(seg_path).change_orientation("RSP")
     seg = resample_nib(seg, new_size=[1, 1, 1], new_size_type="mm", interpolation="nn")
     seg_tensor_all = torch.from_numpy(seg.data.copy())
 
-    img2_path = "/home/GRAMES.POLYMTL.CA/p118739/data_nvme_p118739/data/datasets/spider-challenge-2023/sub-002/anat/sub-002_acq-lowresSag_T2w.nii.gz"
+    img2_path = "/home/ge.polymtl.ca/p118739/data/datasets/spider-challenge-2023/sub-002/anat/sub-002_acq-lowresSag_T2w.nii.gz"
     img2 = Image(img2_path).change_orientation("RSP")
     img2 = resample_nib(img2, new_size=[1, 1, 1], new_size_type="mm", interpolation="linear")
     img2_tensor = torch.from_numpy(img2.data.copy()).to(torch.float32)
 
-    seg2_path = "/home/GRAMES.POLYMTL.CA/p118739/data_nvme_p118739/data/datasets/spider-challenge-2023/derivatives/labels/sub-002/anat/sub-002_acq-lowresSag_T2w_label-spine_dseg.nii.gz"
+    seg2_path = "/home/ge.polymtl.ca/p118739/data/datasets/spider-challenge-2023/derivatives/labels/sub-002/anat/sub-002_acq-lowresSag_T2w_label-spine_dseg.nii.gz"
     seg2 = Image(seg2_path).change_orientation("RSP")
     seg2 = resample_nib(seg2, new_size=[1, 1, 1], new_size_type="mm", interpolation="nn")
     seg2_tensor_all = torch.from_numpy(seg2.data.copy())
@@ -840,9 +580,9 @@ if __name__ == "__main__":
     seg_tensor = torch.cat([seg_tensor, seg2_tensor], dim=0)
 
     # Move to GPU
-    img_tensor = img_tensor.cuda()
-    seg_tensor = seg_tensor.cuda()
-    augmentor = augmentor.cuda()
+    img_tensor = img_tensor.cuda(device=7)
+    seg_tensor = seg_tensor.cuda(device=7)
+    augmentor = augmentor.cuda(device=7)
 
     # Apply augmentations
     augmented_img, augmented_seg = augmentor(img_tensor.clone(), seg_tensor.clone())
@@ -964,5 +704,8 @@ if __name__ == "__main__":
     )
     combined_img2 = np.concatenate([img_line2, seg_line2, augmented_img_line2, augmented_seg_line2, not_augmented_channel_line2], axis=0)
     cv2.imwrite("img/combined2.png", combined_img2 * 255)
+
+    # cv2.imwrite('img/orig_img.png', normalize(pad_numpy_array(img_tensor_np[1, 0, img_tensor_np.shape[2] // 2], pad_shape))*255)
+    # cv2.imwrite('img/aug_img.png', normalize(pad_numpy_array(augmented_img_np[1, 0, img_tensor_np.shape[2] // 2], pad_shape))*255)
 
     print(augmentor)
