+++ smauglab/transforms/synthseg/__init__.py
@@ -1,4 +1,4 @@
-"""SynthSeg generative augmentation for AugLab.
+"""SynthSeg generative augmentation for SmaugLab.
 
 A faithful torch re-implementation of the SynthSeg "brain generator"
 (Billot et al., Medical Image Analysis 2023; BBillot/SynthSeg, BBillot/lab2im):
@@ -9,15 +9,15 @@ Public API:
     SynthSegGenerator      -- the full generative model as an nn.Module
                               (``forward(label_map) -> (image, label)``).
     SynthSegTransformsGPU  -- config-driven driver, ``forward(data, target) ->
-                              (image, target)`` (AugLab calling convention).
+                              (image, target)`` (SmaugLab calling convention).
     RandomSynthSegGPU      -- ImageOnlyTransform that replaces the image with a
                               GMM synthesis of ``params['seg']`` (composes inside
                               AugmentationSequentialCustom pipelines).
 """
 
-from auglab.transforms.synthseg import functional
-from auglab.transforms.synthseg.generator import SynthSegGenerator
-from auglab.transforms.synthseg.transforms import RandomSynthSegGPU, SynthSegTransformsGPU
+from smauglab.transforms.synthseg import functional
+from smauglab.transforms.synthseg.generator import SynthSegGenerator
+from smauglab.transforms.synthseg.transforms import RandomSynthSegGPU, SynthSegTransformsGPU
 
 __all__ = [
     "RandomSynthSegGPU",
